requirements.txt
flask
google-generativeai
pyngrok